requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
dj-database-url==2.3.0
django==5.1.5; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.4.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
pyjwt==2.10.1; python_version >= '3.9'
pytelegrambotapi==4.26.0; python_version >= '3.9'
python-decouple==3.8
requests==2.32.3; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
